Major performance improvement to search preset filters. No longer reads presets from disk on each refresh.
