+++ Solutions/SharePoint.Search.QueryTool/SearchQueryTool/MainWindow.xaml
@@ -409,10 +409,10 @@
 
                                         <StackPanel Margin="5,5,5,0">
                                             <Label Content="Show presets containing:" Margin="5,0,5,0" HorizontalAlignment="Left" VerticalAlignment="Top"/>
-                                            <TextBox x:Name="PresetFilterTextBox" KeyUp="TextBox_KeyEnterUpdate" Margin="5,0,5,0" >
+                                            <TextBox x:Name="PresetFilterTextBox" Margin="5,0,5,0" >
                                             </TextBox>
 
-                                            <ComboBox Name="PresetComboBox" Margin="5" SelectionChanged="PresetComboBox_OnSelectionChanged" >
+                                            <ComboBox Name="PresetComboBox" Margin="5" SelectionChanged="PresetComboBox_OnSelectionChanged" DropDownOpened="PresetComboBox_OnDropDownOpened" >
                                                 <ComboBox.ItemTemplate>
                                                     <DataTemplate>
                                                         <StackPanel Orientation="Horizontal">
